Upgraded the landing page a little
