+++ cmd/yagpdb/templates/cp_selectserver.html
@@ -14,10 +14,11 @@
                 <p>You manage this bot through a web control panel, if you need any help or want to contact the owner you can join the support server (There is a link above)</p>
                 <p>I'm working on adding proper documentation and guides but at this moment the documentation is very lacking, and it may seem a bit complicated. As as i said earlier, join the support server if you have any questions, although please do try a little on your own first, my time is limited.</p>
                 {{if .User}}
-                <a class="btn btn-primary" href="/manage/{{.JoinedGuild.ID}}/stats">Click here to start managing it</a>
+                <a class="btn btn-primary" href="/manage/{{.JoinedGuild.ID}}/stats">Click here to start managing it</a><br/>
                 {{else}}
                 <a class="btn btn-primary" href="/login?goto={{urlquery "/manage/" .JoinedGuild.ID "/stats" }}">Click here to login and start managing it</a>
                 {{end}}
+                <br/>
                 <p>For future reference, you can log in and access your server in the top right menu. if the server did not show up as added in the menu above now, try reloading the page, and if it still does not show up as added then join the support server. (Link above)</p>
             </div>
         </div>
