+++ common/util.go
@@ -411,3 +411,25 @@ func DiscordError(err error) (code int, msg string) {
 
 	return 0, ""
 }
+
+type LoggedExecutedCommand struct {
+	SmallModel
+
+	UserID    string
+	ChannelID string
+	GuildID   string
+
+	// Name of command that was triggered
+	Command string
+	// Raw command with arguments passed
+	RawCommand string
+	// If command returned any error this will be no-empty
+	Error string
+
+	TimeStamp    time.Time
+	ResponseTime int64
+}
+
+func (l LoggedExecutedCommand) TableName() string {
+	return "executed_commands"
+}
