+++ web/web.go
@@ -155,7 +155,7 @@ func setupRoutes() *goji.Mux {
 	mux.Use(UserInfoMiddleware)
 
 	// General handlers
-	mux.Handle(pat.Get("/"), RenderHandler(nil, "index"))
+	mux.Handle(pat.Get("/"), ControllerHandler(HandleLandingPage, "index"))
 	mux.HandleFunc(pat.Get("/login"), HandleLogin)
 	mux.HandleFunc(pat.Get("/confirm_login"), HandleConfirmLogin)
 	mux.HandleFunc(pat.Get("/logout"), HandleLogout)
